Create feature `derive` in `tokio-postgres`. Enabling this feature derives `FromRow` for Rust types. Feature `derive` is not created in `postgres`.
