+++ postgres-derive-test/src/lib.rs
@@ -7,7 +7,6 @@ use std::fmt;
 mod composites;
 mod domains;
 mod enums;
-mod from_row;
 mod transparent;
 
 pub fn test_type<T, S>(conn: &mut Client, sql_type: &str, checks: &[(T, S)])
