+++ tokio-postgres-derive-test/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tokio-postgres-derive-test"
+version = "0.1.0"
+authors = ["tandemdrive"]
+edition = "2018"
+
+[dev-dependencies]
+trybuild = "1.0"
+futures-util = { version = "0.3", features = ["sink"] }
+tokio = { version = "1.0", features = [
+    "macros",
+    "net",
+    "rt",
+    "rt-multi-thread",
+    "time",
+] }
+tokio-postgres = { path = "../tokio-postgres", features = ["derive"]}
