+++ tokio-postgres-derive-test/src/lib.rs
@@ -1,0 +1,3 @@
+#![cfg(test)]
+
+mod from_row;
