+++ tokio-postgres-derive/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "postgres-derive"
-version = "0.4.5"
-authors = ["Steven Fackler <sfackler@palantir.com>"]
+name = "tokio-postgres-derive"
+version = "0.0.0"
+authors = ["tandemdrive"]
 license = "MIT/Apache-2.0"
 edition = "2018"
-description = "An internal crate used by postgres-types"
+description = "Internal proc-macro crate for tokio-postgres"
 repository = "https://github.com/sfackler/rust-postgres"
 
 [lib]
@@ -12,6 +12,7 @@ proc-macro = true
 test = false
 
 [dependencies]
+darling = "0.20"
 syn = "2.0"
 proc-macro2 = "1.0"
 quote = "1.0"
