+++ tokio-postgres/Cargo.toml
@@ -27,6 +27,8 @@ circle-ci = { repository = "sfackler/rust-postgres" }
 default = ["runtime"]
 runtime = ["tokio/net", "tokio/time"]
 
+derive = ["tokio-postgres-derive"]
+
 array-impls = ["postgres-types/array-impls"]
 with-bit-vec-0_6 = ["postgres-types/with-bit-vec-0_6"]
 with-chrono-0_4 = ["postgres-types/with-chrono-0_4"]
@@ -57,6 +59,7 @@ phf = "0.11"
 postgres-protocol = { version = "0.6.6", path = "../postgres-protocol" }
 postgres-types = { version = "0.2.5", path = "../postgres-types" }
 tokio = { version = "1.27", features = ["io-util"] }
+tokio-postgres-derive = { version = "0.0.0", optional = true, path = "../tokio-postgres-derive" }
 tokio-util = { version = "0.7", features = ["codec"] }
 rand = "0.8.5"
 whoami = "1.4.1"
