+++ tokio-postgres/tests/test/types/chrono_04.rs
@@ -1,4 +1,4 @@
-use chrono_04::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, TimeZone, Utc};
+use chrono_04::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, Utc};
 use std::fmt;
 use tokio_postgres::types::{Date, FromSqlOwned, Timestamp};
 use tokio_postgres::Client;
@@ -54,8 +54,9 @@ async fn test_date_time_params() {
     fn make_check(time: &str) -> (Option<DateTime<Utc>>, &str) {
         (
             Some(
-                Utc.datetime_from_str(time, "'%Y-%m-%d %H:%M:%S.%f'")
-                    .unwrap(),
+                DateTime::parse_from_str(time, "'%Y-%m-%d %H:%M:%S.%f %z'")
+                    .unwrap()
+                    .with_timezone(&Utc),
             ),
             time,
         )
@@ -63,9 +64,9 @@ async fn test_date_time_params() {
     test_type(
         "TIMESTAMP WITH TIME ZONE",
         &[
-            make_check("'1970-01-01 00:00:00.010000000'"),
-            make_check("'1965-09-25 11:19:33.100314000'"),
-            make_check("'2010-02-09 23:11:45.120200000'"),
+            make_check("'1970-01-01 00:00:00.010000000 +0000'"),
+            make_check("'1965-09-25 11:19:33.100314000 +0000'"),
+            make_check("'2010-02-09 23:11:45.120200000 +0000'"),
             (None, "NULL"),
         ],
     )
@@ -77,8 +78,9 @@ async fn test_with_special_date_time_params() {
     fn make_check(time: &str) -> (Timestamp<DateTime<Utc>>, &str) {
         (
             Timestamp::Value(
-                Utc.datetime_from_str(time, "'%Y-%m-%d %H:%M:%S.%f'")
-                    .unwrap(),
+                DateTime::parse_from_str(time, "'%Y-%m-%d %H:%M:%S.%f %z'")
+                    .unwrap()
+                    .with_timezone(&Utc),
             ),
             time,
         )
@@ -86,9 +88,9 @@ async fn test_with_special_date_time_params() {
     test_type(
         "TIMESTAMP WITH TIME ZONE",
         &[
-            make_check("'1970-01-01 00:00:00.010000000'"),
-            make_check("'1965-09-25 11:19:33.100314000'"),
-            make_check("'2010-02-09 23:11:45.120200000'"),
+            make_check("'1970-01-01 00:00:00.010000000 +0000'"),
+            make_check("'1965-09-25 11:19:33.100314000 +0000'"),
+            make_check("'2010-02-09 23:11:45.120200000 +0000'"),
             (Timestamp::PosInfinity, "'infinity'"),
             (Timestamp::NegInfinity, "'-infinity'"),
         ],
